+++ ChangeLog.md
@@ -1,6 +1,6 @@
 # Release Notes
-💪 Enhancements
-- GenericEntityReferences support for JsonEntity
 
 🐞 Bug Fixes
-- Fixed support of custom column names in SoftDelete filter for one-to-may relations
+- Number component does not allow saving null value
+- Fix export forms problem
+- Fix job's deactivation